refactor: modulariza tela de checkout
Forms da tela de checkout foram movidos cada um para o seu próprio arquivo
