+++ src/components/Separator.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function Separator() {
+  return <div className="w-full border-gray-500 border-b my-4"></div>;
+}
+
+export default Separator;
