+++ src/forms/AddressForm.js
@@ -1,0 +1,87 @@
+import React from 'react';
+
+import Button from '../components/Button';
+import TextInput from '../components/TextInput';
+import { useForm } from '../context/FormContext';
+import { masks } from '../services/maskService';
+
+const AddressForm = () => {
+  const { handleChange, handleBlur, values, errors, touched, submitForm } = useForm();
+  return (
+    <>
+      <TextInput
+        id="zipCode"
+        placeholder="00000-000"
+        label="CEP"
+        maxLength={9}
+        value={values.zipCode}
+        formatText={current => masks.cep(current, values.zipCode)}
+        onChange={handleChange('zipCode')}
+        onBlur={handleBlur('zipCode')}
+        error={touched['zipCode'] ? errors['zipCode'] : null}
+      />
+      <div className="w-full flex flex-wrap">
+        <div className="w-full sm:w-2/3 sm:pr-4">
+          <TextInput
+            id="street"
+            placeholder="Street"
+            label="Street"
+            value={values.street}
+            onChange={handleChange('street')}
+            onBlur={handleBlur('street')}
+            error={touched['street'] ? errors['street'] : null}
+          />
+        </div>
+        <div className="w-full sm:w-1/3">
+          <TextInput
+            id="number"
+            placeholder="0000"
+            label="Number"
+            value={values.number}
+            onChange={handleChange('number')}
+            onBlur={handleBlur('number')}
+            error={touched['number'] ? errors['number'] : null}
+          />
+        </div>
+      </div>
+      <TextInput
+        id="neighborhood"
+        placeholder="Neighborhood"
+        label="Neighborhood"
+        value={values.neighborhood}
+        onChange={handleChange('neighborhood')}
+        onBlur={handleBlur('neighborhood')}
+        error={touched['neighborhood'] ? errors['neighborhood'] : null}
+      />
+
+      <div className="w-full flex flex-wrap">
+        <div className="w-full sm:w-2/3 sm:pr-4">
+          <TextInput
+            id="city"
+            placeholder="City"
+            label="City"
+            value={values.city}
+            onChange={handleChange('city')}
+            onBlur={handleBlur('city')}
+            error={touched['city'] ? errors['city'] : null}
+          />
+        </div>
+        <div className="w-full sm:w-1/3">
+          <TextInput
+            id="state"
+            placeholder="State"
+            label="State"
+            value={values.state}
+            onChange={handleChange('state')}
+            onBlur={handleBlur('state')}
+            error={touched['state'] ? errors['state'] : null}
+          />
+        </div>
+      </div>
+      <div className="w-full mt-12">
+        <Button text="Next" onClick={submitForm} />
+      </div>
+    </>
+  );
+};
+export default AddressForm;
